requirements.txt
Flask
replit
psycopg2-binary
